+++ BodoSQL/bodosql/tests/conftest.py
@@ -21,20 +21,12 @@ from pyspark.sql.types import (
 import bodo
 import bodo.utils.allocation_tracking
 import bodosql
-from bodo.io.iceberg.catalog import conn_str_to_catalog
 from bodo.tests.conftest import (  # noqa
-    aws_polaris_warehouse,
-    azure_polaris_warehouse,
     iceberg_database,
     memory_leak_check,
-    polaris_connection,
-    polaris_package,
-    polaris_server,
-    polaris_token,
 )
 from bodo.tests.iceberg_database_helpers.utils import get_spark
-from bodo.tests.utils import gen_nonascii_list, get_rest_catalog_connection_string
-from bodo.utils.utils import run_rank0
+from bodo.tests.utils import gen_nonascii_list
 
 # Patch to avoid PySpark's Py4j exception handler in testing.
 # See:
@@ -1956,45 +1948,6 @@ def listagg_data():
             }
         )
     }
-
-
-@pytest.fixture()
-def polaris_catalog_iceberg_read_df(polaris_connection):
-    rest_uri, polaris_warehouse, polaris_credential = polaris_connection
-    df = pd.DataFrame(
-        {
-            "A": pd.array(["ally", "bob", "cassie", "david", pd.NA]),
-            "B": pd.array([10.5, -124.0, 11.11, 456.2, -8e2], dtype="float64"),
-            "C": pd.array([True, pd.NA, False, pd.NA, pd.NA], dtype="boolean"),
-        }
-    )
-    con_str = get_rest_catalog_connection_string(
-        rest_uri, polaris_warehouse, polaris_credential
-    )
-    py_catalog = conn_str_to_catalog(con_str)
-    table_id = "CI.BODOSQL_ICEBERG_READ_TEST"
-    run_rank0(
-        lambda: py_catalog.create_table(table_id, pa.Schema.from_pandas(df)).append(
-            pa.Table.from_pandas(df)
-        )
-    )()
-    yield df
-    run_rank0(lambda: py_catalog.purge_table(table_id))()
-
-
-@pytest.fixture
-def polaris_catalog(polaris_connection):
-    """
-    Returns a polaris catalog object
-    """
-    rest_uri, polaris_warehouse, polaris_credential = polaris_connection
-
-    return bodosql.RESTCatalog(
-        rest_uri=rest_uri,
-        warehouse=polaris_warehouse,
-        credential=polaris_credential,
-        scope="PRINCIPAL_ROLE:ALL",
-    )
 
 
 @pytest.fixture
