+++ BodoSQL/bodosql/tests/test_types/test_rest_iceberg_catalog.py
@@ -9,7 +9,11 @@ import pytest
 import bodo
 import bodosql
 from bodo.io.iceberg.catalog import conn_str_to_catalog
-from bodo.tests.iceberg_database_helpers.utils import SparkAwsIcebergCatalog, get_spark
+from bodo.tests.iceberg_database_helpers.utils import (
+    SparkAwsIcebergCatalog,
+    create_iceberg_table,
+    get_spark,
+)
 from bodo.tests.user_logging_utils import (
     check_logger_msg,
     create_string_io_logger,
@@ -18,7 +22,6 @@ from bodo.tests.user_logging_utils import (
 from bodo.tests.utils import (
     assert_tables_equal,
     check_func,
-    create_polaris_iceberg_table,
     gen_unique_table_id,
     get_rest_catalog_connection_string,
     pytest_polaris,
@@ -594,9 +597,10 @@ def test_dynamic_scalar_filter_pushdown(
     column = [current_date + pd.Timedelta(days=offset) for offset in offsets]
     input_df = pd.DataFrame({"A": column})
     py_output = pd.DataFrame({"A": [x for x in column if x <= current_date]})
-    with create_polaris_iceberg_table(
-        input_df, "current_date_table", conn_str, schema
-    ) as table_name:
+    conn_str_to_catalog(conn_str)
+    table_name = "current_date_table"
+    table_id = f"{schema}.table_name"
+    with create_iceberg_table(conn_str, table_id, input_df):
         query = f'SELECT * FROM {schema}."{table_name}" WHERE A <= CURRENT_DATE'
         stream = StringIO()
         logger = create_string_io_logger(stream)
